refactored CoAP operations into their own module
